#314: rename folders components + remove useless code + add some css rules
