+++ webapp/app/components/Login/styles.css
@@ -1,8 +1,3 @@
-.login {
-  position: absolute;
-  left: 0;
-}
-
 .titre {
   margin:auto;
 }
@@ -11,10 +6,6 @@
   display: block;
   width: 400px;
   margin: auto;
-}
-
-body {
-  overflow: visible;
 }
 
 .form {
@@ -29,4 +20,9 @@ body {
 .toolbar {
   margin-top: 20px;
   text-align: right;
+}
+
+.colLogin {
+  float: none!important;
+  margin: auto;
 }
