+++ webapp/app/containers/Login/Form/constants.js
@@ -3,4 +3,3 @@
 //
 
 export const LOGIN = 'LOGIN';
-export const SET_USER_INFO = 'SET_USER_INFO';
