+++ webapp/app/reducers.js
@@ -13,7 +13,7 @@ import StatsReducer from 'containers/Stats/reducers';
 
 import dashboardReducer from 'containers/Dashboard/reducers';
 import forgottenPasswordReducer from 'containers/ForgottenPassword/Form/Email/reducer';
-import appReducer from 'containers/App/reducers';
+import appReducer from 'containers/App/reducer';
 import loginReducer from 'containers/Login/Form/reducers';
 import createUserReducer from 'containers/CreateUser/Form/reducers';
 import createGroupReducer from 'containers/CreateGroup/Form/reducers';
